agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
google-adk
google-genai
ag-ui-adk
tavily-python
dotenv
fastmcp
mcp
mailtrap
